fixed stuff

+++ software/3D-CMCC-Forest-Model/headers/settings.h
@@ -30,7 +30,7 @@ typedef struct {
 	double init_frac_maxasw; /* minimum fraction of available soil water at the beginning of simulation */
 
 	double tree_layer_limit;
-	int soil_layer;
+	double soil_layer;
 	double min_layer_cover;
 	double max_layer_cover;
